git-tfs-id: [http://hillstfs2013:8080/tfs/Phytel]$/PhytelCode/Phytel.Net/Services;C1371

+++ API/DataDomain/Phytel.API.DataDomain.PatientObservation/Development/Current/PatientObservation/MongoDB/DTO/MEPatientObservation.cs
@@ -38,7 +38,7 @@ namespace Phytel.API.DataDomain.PatientObservation.DTO
         public const string NumericValueProperty = "nval";
         [BsonElement(NumericValueProperty)]
         [BsonIgnoreIfNull(true)]
-        public float? NumericValue { get; set; }
+        public double? NumericValue { get; set; }
 
         public const string NonNumericValueProperty = "nnval";
         [BsonElement(NonNumericValueProperty)]
